No boost dependency anymore
Newer GT4Py versions come with a vendored-in version of the one boost
header that they need. NDSL picked up this newer version and this PR
propagates that change to PySHiELD, which now doesn't need to depend on
boost headers in the `Dockerfile`. Not sure if the `Dockerfile` is still
in use, but it was easy enough to just update it.